requirements.txt
# --- Web Framework ---
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.9

# --- Image Processing ---
pillow>=10.0.0
pillow-heif>=0.10.0
opencv-python-headless>=4.8.0

# --- Face Recognition ---
insightface>=0.7.3
onnxruntime>=1.15.0
numpy>=1.24.0

# --- Scripts 
rawpy>=0.18.0
imageio>=2.30.0
google-api-python-client>=2.90.0
google-auth>=2.20.0
